Revert "Lakshya and jai"
