+++ Database Project 1/src/BTree_search.cpp
@@ -6,57 +6,60 @@
 
 using namespace std; 
 
- std::vector<Key_Records> Btree::search(BTreeNode* node, int key, bool rangeflag, int key2){
+ std::vector<Key_Records> Btree::search(int key, bool rangeflag, int key2){
 
     std::vector<Key_Records> search_result = {};
 
     //Tree is empty
-    if(node == nullptr) {
+    if(root == nullptr) {
         cout<<"B+ Tree is empty\n";
         return;
     }
 
-    //Traverse internal nodes 
-    if(!node->isleaf){
-        for(int i=0 ; i<node->numKeysPerNode;++i){
-            if (key < node->keys->key_value) {
-                return search(node->child[i], key, rangeflag, key2);
-            }
-            
-            //reach the end of the keys in the node 
-            if (key == node->numKeysPerNode - 1) {
-                return search(node->child[i+1], key, rangeflag, key2);
-            }
-        }        
-    }
+    else{
+        BTreeNode* current = root;
 
-    else { //leaf node 
+        //traverse through internal nodes 
+        while(current->isleaf == false){
+            for(int i=0; i<current->numKeysPerNode; i++){
+                if(key < current->keys[i].key_value){
+                    current = current->child[i]; 
+                    break;
+                }
 
-        //If not a range query
-        if(!rangeflag){
-            for (int i = 0; i < node->numKeysPerNode; ++i) {
-                if (node->keys->key_value == key) {
-                cout<<"Key Found!\n";
-                search_result.push_back(node->keys[i]);
-                return search_result; 
+                //if we reach the end of the keys in the node
+                if(key == current->numKeysPerNode - 1){
+                    current = current->child[i+1];
+                    break;
                 }
             }
         }
-        
-        //Range query
-        else{
-            for (int i = 0; i < node->numKeysPerNode; ++i) {
-                if(node->keys->key_value >= key && node->keys->key_value <= key2){
-                    search_result.push_back(node->keys[i]);
 
-                    if(i==node->numKeysPerNode-1){
-                        node = node->child[i];
+        //leaf node 
+        if(!rangeflag){ //Single key query
+            for(int i = 0; i<current->numKeysPerNode; i++){
+                if (current->keys[i].key_value == key) {
+                    cout<<"Key Found!\n";
+                    search_result.push_back(current->keys[i]);
+                    return search_result;
+                }
+            }
+        }
+
+        else{ //Range query
+            for (int i = 0; i < current->numKeysPerNode; ++i) {
+                if(current->keys[i].key_value >= key && current->keys[i].key_value <= key2){
+                    search_result.push_back(current->keys[i]);
+
+                    if(i==current->numKeysPerNode-1){
+                        current = current->child[i];
                         i = 0;
                     }
                 }
             }
             return search_result;
         }
+
     }
     return search_result;
 };
